refactor(encode): 重构加密算法模块, JwtUtil 类
