+++ xingtools-core/src/main/java/top/cutexingluo/tools/utils/ee/token/JwtUtil.java
@@ -7,8 +7,6 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 
 import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import java.util.Base64;
 import java.util.Date;
 import java.util.UUID;
 
@@ -21,34 +19,26 @@ import java.util.UUID;
  */
 public class JwtUtil {
 
-    //有效期为
+
+    /**
+     * 签发者
+     */
+    public static String ISSUER;
+
+    /**
+     * 有效期为 (毫秒)
+     */
     public static Long JWT_TTL = 24 * 60 * 60 * 1000L;// 60 * 60 *1000  一个小时
-    //设置秘钥明文
-    public static String JWT_KEY = "xingtian";
-
-    // 签发者
-    public static String ISSUER = "xt";
 
     /**
-     * 设置默认有效时间
+     * 秘钥对象
      */
-    public static void setJWTTtl(long jwtTtl) {
-        JwtUtil.JWT_TTL = jwtTtl;
-    }
+    public static SecretKey SECRET_KEY;
 
     /**
-     * 设置JWT key
+     * 加签算法 (默认 HS256)
      */
-    public static void setJWTKey(String jwtKey) {
-        JwtUtil.JWT_KEY = jwtKey;
-    }
-
-    /**
-     * 设置签发者
-     */
-    public static void setISSUER(String issuer) {
-        JwtUtil.ISSUER = issuer;
-    }
+    public static SignatureAlgorithm SIGNATURE_ALGORITHM = SignatureAlgorithm.HS256;
 
     /**
      * 无分隔符 "-" 的UUID
@@ -80,8 +70,8 @@ public class JwtUtil {
     }
 
     private static JwtBuilder getJwtBuilder(String subject, Long ttlMillis, String uuid) {
-        SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS256;
-        SecretKey secretKey = generalKey();
+        SignatureAlgorithm signatureAlgorithm = SIGNATURE_ALGORITHM;
+        SecretKey secretKey = SECRET_KEY;
         long nowMillis = System.currentTimeMillis();
         Date now = new Date(nowMillis);
         if (ttlMillis == null) {
@@ -89,13 +79,18 @@ public class JwtUtil {
         }
         long expMillis = nowMillis + ttlMillis;
         Date expDate = new Date(expMillis);
-        return Jwts.builder()
+        JwtBuilder jwtBuilder = Jwts.builder()
                 .setId(uuid)              //唯一的ID
                 .setSubject(subject)   // 主题  可以是JSON数据
-                .setIssuer(ISSUER)     // 签发者
                 .setIssuedAt(now)      // 签发时间
-                .signWith(signatureAlgorithm, secretKey) //使用HS256对称加密算法签名, 第二个参数为秘钥
                 .setExpiration(expDate);
+        if (ISSUER != null) {
+            jwtBuilder.setIssuer(ISSUER);     // 签发者
+        }
+        if (secretKey != null) {
+            jwtBuilder.signWith(signatureAlgorithm, secretKey); //使用算法签名, 第二个参数为秘钥
+        }
+        return jwtBuilder;
     }
 
     /**
@@ -112,20 +107,12 @@ public class JwtUtil {
         System.out.println(claims);
     }
 
-    /**
-     * 生成加密后的秘钥 secretKey
-     */
-    public static SecretKey generalKey() {
-        byte[] encodedKey = Base64.getDecoder().decode(JwtUtil.JWT_KEY);
-        SecretKey key = new SecretKeySpec(encodedKey, 0, encodedKey.length, "AES");
-        return key;
-    }
 
     /**
      * 解析
      */
     public static Claims parseJWT(String jwt) throws Exception {
-        SecretKey secretKey = generalKey();
+        SecretKey secretKey = SECRET_KEY;
         return Jwts.parser()
                 .setSigningKey(secretKey)
                 .parseClaimsJws(jwt)
